+++ packages/eslint-plugin/src/Elements/Elements.spec.ts
@@ -13,7 +13,8 @@ const createSettings = (
     rootPath: "/repo",
     dependencyNodes: [],
     legacyTemplates: false,
-    elementsSingleType: false,
+    elementsSingleMatch: false,
+    filesSingleMatch: false,
     cache: true,
     flagAsExternal: {},
     debug: {